requirements.txt
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
idna==3.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pdfminer==20191125
pip-check-reqs==2.3.2
platformdirs==2.5.1
playsound==1.2.2
pycryptodome==3.14.1
pyparsing==3.0.7
python-dotenv==0.20.0
requests==2.27.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
